+++ serpentTools/objects/containers.py
@@ -7,14 +7,13 @@ Contents
 * :py:class:`~serpentTools.objects.containers.BranchContainer`
 
 """
-from re import compile
 from itertools import product
 from warnings import warn
 from numbers import Real
 
 from six import iteritems, PY2
 from matplotlib import pyplot
-from numpy import array, arange, hstack, ndarray, zeros_like
+from numpy import arange, hstack, ndarray, zeros_like
 
 from serpentTools.settings import rc
 from serpentTools.objects.base import NamedObject, BaseObject
@@ -61,9 +60,6 @@ HOMOG_VAR_TO_ATTR = {
 __all__ = (
     'HomogUniv', 'BranchContainer',
 )
-
-
-CRIT_RE = compile('K[EI][NF]F')
 
 
 class HomogUniv(NamedObject):
@@ -241,14 +237,9 @@ class HomogUniv(NamedObject):
         """
         Return the new value to be stored after some cleaning.
 
-        Makes sure all vectors, everything but keff/kinf data, are
-        converted to numpy arrays. Reshapes scattering matrices
-        if number of groups is known and ``xs.reshapeScatter``
+        Reshapes scattering matrices if number of groups
+        is known and ``xs.reshapeScatter``
         """
-        if not isinstance(value, ndarray):
-            value = array(value)
-        if CRIT_RE.search(name):
-            return value[0]
         ng = self.numGroups
         if self.__reshaped and name in SCATTER_MATS:
             if ng is None: